+++ lib/src/types/bson_regexp.dart
@@ -60,7 +60,8 @@ class BsonRegexp extends BsonObject {
       throw ArgumentError(
           'The received Map is not a avalid EJson Regex representation');
     }
-    if (entry.value is! Map<String, Object>) {
+    if (entry.value['pattern'] is! String ||
+        entry.value['options'] is! String) {
       throw ArgumentError(
           'The received Map is not a valid EJson Regex representation');
     }